Conflicts:
	Libs/XNAT/Core/ctkXnatDefaultSchemaTypes.cpp
	Libs/XNAT/Core/ctkXnatDefaultSchemaTypes.h
	Libs/XNAT/Core/ctkXnatExperiment.cpp
	Libs/XNAT/Core/ctkXnatSession.cpp

+++ Libs/Core/ctkBackTrace.cpp
@@ -178,7 +178,7 @@ int ctkBackTracePrivate::trace(void** array, size_t n) const
 #else
 
 // --------------------------------------------------------------------------
-int ctkBackTracePrivate::trace(void** /*array*/, size_t /*n*/) const
+int ctkBackTracePrivate::trace(void** /*array*/, int /*n*/) const
 {
   return 0;
 }
